Update stackviz tarball location
We've patched stackviz to work under python3 properly but we are still
pulling an old tarball for stackviz that was built last year. The legacy
job that built the file at this location seems to have been removed.
Switch to the new dist/ location which appears to be correct based on
tarball file sizes.

Someone that understands stackviz better than me should confirm this new
locations is the correct one.

Change-Id: If659a6f1fb50d288afed75e3f4975f7a4d140d35