New Airship project - Tempest Plugin
This PS adds a Tempest plugin project for Airship, which will
be used to confirm the correctness of APIs for Airship projects.
This is being created as a standalone project per OpenStack
conventions.

The existing airship-tempest-plugin project to be imported contains
a subset of tests for the Shipyard project; the tests will expand
over time to test other Airship projects as well.

Change-Id: I58d09f41d7f696de628d42cc5504572811d38c61