wheel-mirror: serialise copies to AFS
This is a follow-on to I3761dd9d2ab752e66bcccdd2434405f11535b95f where
I misunderstood what "linear" was doing.  What was really intended was
to stop both the python2 and python3 builds copying to AFS at the same
time.  "linear" just means the *tasks* start at the same time across
hosts; not that they run serially.  To do that you want the (aptly
named) "serial" specifier on the playbook.

Add this so we do the copies sequentially

Change-Id: I4f8a2f2c6c8164e7ea207f8e4b286e06df0b13dd