Revert "Disable ovh-gra1"
This partially reverts commit
bfdd3e6a4234747ce637a8ef007821f90caac030.

After fruitful discussions with amorin in IRC, we have nodes working
again in this region.  This puts a small load on for us to monitor for
a while.  A follow-on will do a full revert so we don't forget.

Story: #2004090
Task: #27492
Change-Id: Id01f85fcee150f9360f508b09003a8d0043155bd