Update ansible library location to envsitepackagesdir
We install zuul so that we can pick up library entries. The proper path
for this is envsitepackagesdir. Also, just list zuul in the requirements
so that we're not cloning remotely. tox-siblings will take care of it in
the gate for us, because we add zuul to required-projects in the
openstack-zuul-jobs-linters job.

This adds a setup.py and setup.cfg so that tox-siblings can do its
thing.

Depends-On: https://review.openstack.org/611607
Change-Id: I864308af3424a0b0505a252d0d5cf44277172fbf