wheel-build : temporary pip with checksum output
We are currently trying to debug persistent wheel corruption.  So far,
we have been completely unable to replicate the corrupt output outside
the periodic jobs.

A first thought was that it was due to AFS corruption due to multiple
writes (I4f8a2f2c6c8164e7ea207f8e4b286e06df0b13dd), however this does
not appear to be the case.

Experimentally manually running the builds, on infra nodes, under
python2 and python3, in parallel just as done here, does not replicate
the problem.

This patched version of pip will output the sha256 hash for the final
build output of each wheel.  The plan is to correlate that against any
corrupt file that appears.  If the corrupt hash matches any file
produced by pip, then we know the problem is inside pip (and we will
have the exact build situation that it occured in); if the corrupt
file does not match then we must have some sort of issue copying the
files or similar.

Change-Id: I81943ed459bf4e2c77cae42e50af5fc5979682b4