Switch nodepool builders to zk cluster
Switch over the nodepool builders to our newer zk cluster from the old
single node zk cluster. We will stop building images that the launchers
can see before the launchers move, but this lets us preseed the new
cluster with up to date image data.

Once the images are built with records in the new zk cluster we can
switch over the zuul scheduler and the launchers to this newer cluster.

Change-Id: I95ca326095decc03cf279383fa48dbdfc56ed8c8