Use python3 for release
Update the release-openstack-python job to use python3, it thus will
be the same as release-openstack-python3 which we can remove in a
followup change.

There's no need to have two variants, we can safely use the python3
variant everywhere.

Add also a test-release-openstack job as duplicate of
test-release-openstack-python3.

The goal of this whole patch series for project-config and
openstack-zuul-jobs is to give publish-to-pypi exactly the same set up
as publish-to-pypi-python3 has - and then use the updated
publish-to-pypi template everywhere so that we can remove
publish-to-pypi-python3.

Needed-By: https://review.openstack.org/615237
Change-Id: I09cf4e64194395417c1bfa95237b04bd316c18fa