Neutron grafana update for co-gating section
The networking-ovn-tempest-dsvm-ovs-release job needed to be added to
the neutron grafana dashboard, as part of a new initiative to allow
stadium projects to add a non-voting job to make sure there is not
breakage as neutron master changes.  Since ironic already has a job
doing this, consolidate them into a new section for co-gating jobs.

Also add the tempest-slow job to the gating jobs, as specified in the
change upon which this depends.

Depends-On: https://review.openstack.org/622419
Change-Id: I20605bef23218a06d04b188a5829967a2a7d8e5c