Halve bhs1 max-servers value
We are debugging slow nodes in bhs1. Looking at dstat data we clearly
have some jobs that end up spending a lot of cpu time in sys and wai
columns while other similar jobs do not.

One thought was that this is due to an unhappy hypervisor or two, but
amorin has dug in and found that these slow jobs run on multiple unique
hypervisors implying that isn't likely.

My next thought is that we are our own noisy neighbors. Reducing the
max-servers should improve things if we are indeed our own noisy
neighbors.

Change-Id: Idd7804778a141d38da38b739294c6c6a62016053