Change os-resource-classes and os-traits acl config to placement
During the initial creation of the os-resource-classes repo it
was put under nova.config. This was probably not the right choice
but was done to be aligned with os-traits, which is a very
similar library.

Instead of keeping os-resource-classes in the past for the sake of
os-traits, let's move them both to the future, and include
osc-placement, the openstack client plugin for placement, as well.

The placement.config is a superset of the nova.config. If/when
placement is its own project team, os-resource-classes, os-traits,
and osc-placement will be members of the deliverables.

Change-Id: Ice1b0476bfb7471ddab836df44dd26eaba8ccd19