Use OVH BHS1 again
Set ovh-bhs1 max-servers to 150. OVH (thank you amorin) have debugged
and corrected a memory leak there that we believe to be the cause of the
test node slowness.

Frickler and I have run fio tests on VMs running on each hypervisor in
the region and they look happy. We've also run spot tests of devstack
and tempest which also appear happy.

Change-Id: If6fd5a6194a9996e8b031f74918f373dc7bbe758