kata: add dco-license jobs to kata projects
Now the dco-license check is part of the baseline zuul
library, add it as a pre-check (dependency) to the kata
projects.

We make it a pre-check, as if the signoff is incorrect,
the whole job will be re-submitted again, so why build
and test it all if we know we are going to have to
do it again later anyway. Sure, maybe a build will catch
problems 'earlier', but much of the time the submitter
may either fix the signoff before the builds have even
completed, or just see the signoff error and re-submit
without checking the rest of the build logs anyway.

Change-Id: I75c31b63bc4db3160162aac69b8d4bf94e32a942
Signed-off-by: Graham Whaley <graham.whaley@intel.com>