Run puppetforge publish playbook in the clouds
Changes the hosts field from localhost to all
so that the publish playbook is executed in
the cloud instead of locally on the zuul executor
node.

This is required because when it's executed on the
zuul executor the security context supresses the
facts as a security measure which causes the job to fail.

Change-Id: Iab2252ca647bd6d8eaff6c48f458c566c0393cb2