Remove some base jobs
These are now in opendev/base-jobs and can be removed from this repo.

The site_logs secret is defined in opendev/base-jobs, so for the cases
where we are using it to publish things other than logs, duplicate it
and rename it.  A couple of base jobs remain and use it as their logs
secret, so rename it for those cases, until we remove them to
opendev/base-jobs as well.

Change-Id: If3d281c657f50b82ec57fe8649fef6653b7ceec5