Fix translation jobs due to opendev/base-job change
These jobs uses a role in openstack-zuul-jobs, which is a repo that
used to be included in openstack's base job, but is no longer
now that we switched to the opendev base job.  This adds the repo
back to the role path for this job.

Failure is:
ERROR! the role 'prepare-zanata-client' was not found...

Change-Id: Id23965a2d43e555fe1f6c33de1381813fdf7239c