Enable DIB_SHOW_IMAGE_USAGE for build
This prints a report in the dib output logs of files and directories
over 10MiB in size.

There is some history behind this.  It was originally enabled in
nodepool for all builds, but disabled after we started seeing hangs
reading from pipes [1,2].

A lot has changed since then.  This was Python 2 on trusty hosts.  We
are now python 3 on Xenial hosts and the logging capture has been
rewritten [3].  This is also enabled in our gate tests for all builds
[4] and has not caused problems (though there is much less in the
images).

Note this doesn't actually increase build time -- dib already does the
du to figure out how big to make the loopback image as is.  This
re-enables it for all builds.

[1] https://review.openstack.org/357393
[2] http://eavesdrop.openstack.org/irclogs/%23openstack-infra/%23openstack-infra.2016-08-18.log.html#t2016-08-18T20:49:51
[3] https://review.openstack.org/633792
[4] https://git.openstack.org/cgit/openstack-infra/nodepool/tree/devstack/plugin.sh#n479

Change-Id: Iea8b101d723205b30d4a757eaeadb0dccb7243f5