Remove tox-py35-on-zuul from zuul-jobs
We now have some real usage of tox on this repo itself, so the
self-testing of the tox environment we got from running this job
is now supplied by that.  The tox-py35-on-zuul job takes a while
to run and is sometimes flaky, so remove it so that it doesn't
act as an unecessary blocker.

Change-Id: Ifc274f049381cacead030e4baac460a7697c20cc