Set up placement project to use storyboard
All four of placement, os-traits, os-resource-classes and
osc-placement will be in the placement group.

Note that at this time no migration is being requested.
After discussion in #storyboard [1] along with discussion in
email and IRC with placement team members we've determined
this shouldn't present too much of an issue because:

* os-traits and os-resource-classes currently have no bugs

* within the placement project we've discussed letting at least
  some in progress bugs and blueprints (currently living in the
  nova project) complete there, through to the end of the stein
  cycle

* remaining placement bugs use a 'placement' tag and can
  be migrated in the future via an 'lp bugtag filter'

* osc-placement bugs can also be migrated at a later time when
  we fully define a plan

* updating bug progress for any of these projects, from gerrit,
  is not working for some time now, presumably since the change
  in acls

The 'future' or 'later time' defined plan will come through some
combination of the passage of time and a PTL being elected but
in the meantime we'd like to use storyboard to start remembering
things that need to be done at some point.

[1] http://eavesdrop.openstack.org/irclogs/%23storyboard/%23storyboard.2019-02-26.log.html#t2019-02-26T17:48:26

Change-Id: Idde65c5b5b10897bc96700cf140b56aece94b11f