Sphinx publishing adjustments for ubuntu-bionic
First, the publish-openstack-sphinx-docs-base job is deprecated
since Rocky so should not be used for any branches which test on
ubuntu-bionic nodes. As a result it's reasonable to force it to
always use an ubuntu-xenial nodeset instead.

Second, its replacement publish-openstack-tox-docs-base needs two
variants, one which uses an ubuntu-xenial nodeset on stable/ocata
through stable/rocky and another which uses the default
(ubuntu-bionic) nodeset for stable/stein and later.

Change-Id: If1b93c39ed978d76df4286d52d28af73ea1de574