Add some projects to zuul zuul tenant
This adds zuul-jobs (which is needed for roles and jobs); since it
does not contain its own project stanza, it won't be "active" in
the tenant yet, it will just be present for use in jobs.

It also moves zuul-sphinx which does manage its own jobs.  It is
a slow-moving repo so we can test some jobs on it before we move
the rest of the projects over.

Change-Id: I3c5db1a9f88abce3d92f10dbb1443a987e08648f