Add some helpers to projectconfig_ruamellib
So that it's easier to use in our scripts, add some helper sugar.

First, add functions load and dump so that the library can be used
similarly to yaml without the need to construct an object first.

Then, remove the strip parameter and detect whether we're dumping
a list or not. That way people won't forget to pass or not pass
strip.

Change-Id: I204af8a89c37f36f0480de3a2e669b65354eb73c