Add glance, tempest, and nova stable* to ceph grafana
This adds the glance and tempest projects to the ceph failure rate
tracking dashboard. It also adds the nova stable branches to the view.
Finally, the default time period is changed to the last 7 days and the
panel width increased to look better (idea stolen from the neutron
grafana dashboard).

Change-Id: I0f099a615b511b64f2347dbb322bcf3d29ac2bef