Set min-ready 0 on vexxhost specific labels
It seems these nodes are not used very often, compared to other labels
on vexxhost.  By drop these 3 to zero, it means we can can 3 more nodes
for 99% of the time.  Give that vexxhost is a fast booting clouds, this
shouldn't impact job runtimes too much for tests that use these specific
labels.

Change-Id: I7e6f6e2395448cb241a76de450771c63452a93ec
Signed-off-by: Paul Belanger <pabelanger@redhat.com>