wheel-build: Work around pip 19.0 errors
As pointed out by ajaeger; the wheel build jobs are failing with

 Error installing 'pip from
 git+https://github.com/ianw/pip.git@path-and-hash#egg=pip': editable
 mode is not supported for pyproject.toml-style projects.

This appears to be an issue with pip 19.0 which would be installed on
the build hosts, interacting with projects that have a pyproject.toml
file.

The problem only seems to be related to installing in editable mode,
which we don't really need here on this ephemeral host.  So just
switch to a regular install, with upgrade/reinstall to be sure.

[1] https://github.com/pypa/pip/pull/6449

Change-Id: I006500a2c98f512176fbac4ab426a8719243fc70