Increase nodepool image journal size
DIB uses 64MB as the default ext4 journal size, which is far below
the mkfs.ext4 default for the 80GB disks that nodepool instances
use. This change will increase the journal size to 512MB which is
the default size for ext4 filesystems built on disks 64-128GB in
size.

The relevant mkfs source where default journal sizes are calculated
is available here:
https://git.kernel.org/pub/scm/fs/ext2/e2fsprogs.git/tree/lib/ext2fs/mkjournal.c#333

http://www.linux-mag.com/id/7666/ shows that an undersized journal
can have a large impact on file operations when dealing with lots
of small files. With larger files, there is negligible performance
impact. So worst case this change may not improve performance, but
with the number of jobs using small python source files and small
.git repo files, this should cause a marginal performance increase
for the nodepool use case.

Change-Id: I07a67e2e51d8d139e27da83aa16899a11d8b3f34