Switch bindep fallback testing to openSUSE Leap 15.x
openSUSE Leap 42.3 is end of life in June 2019 so testing
it further doesn't make a whole lot of sense. instead we're
adding openSUSE Leap 15 (referring to openSUSE leap 15.0 at
the moment but once the infrastructure for Leap 15.1 has been
built it will be updated to the leap 15.x nodepool image). The
main reason why we can do this cleanup now is because there
has been a policy shift with openSUSE Leap 15: the minor releases
like 15.1 and 15.2 are similar to centos 7.x releases, they
should be all backward compatible so referring to it as leap15
is the correct thing to do.

Change-Id: I7ec755e678a3dfe45efad9104884df3daf315d2c