Add openSUSE 15.1 to the nodepool building as opensuse-15
The main reason why we can do this cleanup now is because there has been a policy
change with openSUSE Leap 15: the minor releases like 15.1 and 15.2 are similarly
backwards compatible like e.g. minor releases in centos 7.x are, as such
we can build an opensuse 15 image in the ci and update all jobs to that
one to have less continuous effort in maintaining the opensuse builds.

Depends-On: https://review.opendev.org/#/c/660137/

Change-Id: I2b1f21fb6e01558c8cee27de116dfc857a1a1c91