opendev.org mirrors: switch out DFW, switch in IAD
We have identified issues with openafs meaning the DFW opendev.org
mirror is unstable.  The IAD mirror is running a (currently
hand-installed) version of kafs to provide the mirror directories (the
host is currently in the emergency file).

Switch in IAD for testing kafs, and switch out DFW which is giving
occasional errors to jobs.

Change-Id: I5115030cd323db459b248e21413b68f0ad0aafa8