Retire independent ansible role
All of the tripleo ansible roles have been retired. The transformation squad
is pivoting away from independent roles at this time and will be focusing on
the tripleo-ansible repository. For more information on the tripleo-ansible
efforts and the direction of the transformation squad please review the
following mailing list thread [0].

[0] - http://lists.openstack.org/pipermail/openstack-discuss/2019-June/006836.html

Change-Id: Ia22150be91e7f45f59be4bab381727faa3d637df
Signed-off-by: Kevin Carter <kecarter@redhat.com>