Remove nodepool builder config for control plane images
Nodepool wants to build a new image every day then delete the N+2'th
image. This is great for dynamic test nodes which never live for longer
than a few hours. Unfortunately, with boot from volume servers you
cannot delete their base image if the server is still running. This
means as soon as we use these images for control plane servers, which do
not go away after a few hours, we'll no longer be able to delete that
image a couple days later.

This will make our current nodepool install unhappy so lets avoid doing
that for now.

We can probably make improvements to nodepool to better support this use
case. At that time we can start trying this again. We will probably also
want to use a dedicated builder for that.

Note this isn't a revert because the control-plane-minimal element is
useful if we want to build these types of images in the future. We just
shouldn't use nodepool to do it in its current state.

Change-Id: I610dae5d6d7fe01e15f66970b59d3b999bebb766