Stop building control plane images
Nodepool wants to build a new image every day then delete the N+2'th
image. This is great for dynamic test nodes which never live for longer
than a few hours. Unfortunately, with boot from volume servers you
cannot delete their base image if the server is still running. This
means as soon as we use these images for control plane servers, which do
not go away after a few hours, we'll no longer be able to delete that
image a couple days later.

This will make our current nodepool install unhappy so lets avoid doing
that for now.

We can probably make improvements to nodepool to better support this use
case. At that time we can start trying this again. We will probably also
want to use a dedicated builder for that.

Note this isn't a revert because we want to wind down nodepool in such a
way that it cleans up after itself.

Change-Id: Ic45b52f6bfc803756a8f42f21c848e1fd223175f