Retire independent ansible role
All of the tripleo ansible roles have been retired. The transformation squad
is pivoting away from independent roles at this time and will be focusing on
the tripleo-ansible repository. For more information on the tripleo-ansible
efforts and the direction of the transformation squad please review the
following mailing list thread [0].

This change adds zuul project config for the cookiecutter role which was
missing.

[0] - http://lists.openstack.org/pipermail/openstack-discuss/2019-June/006836.html

Change-Id: Ic2d23cb2078d6ef00c13056a0b1c4fbac1d7fdde
Signed-off-by: Kevin Carter <kecarter@redhat.com>