New project request: airship/docs
This adds a documentation project for Airship.  Currently, Airship
documentation is scattered across a handful of non-obvious projects.
At the PTG it was decided to create a single, clear home for
Airship-wide documentation as the project matures.
Existing documentation will migrate to this project, and new docs
will be authored here, once the project is created.

Change-Id: Ifab77cc746e0187611eda98776715909e4819af8