Load extra Zuul config from zuul-jobs
In the "zuul" Zuul tenant only, load extra Zuul config from the
zuul-tests.d/ directory of the zuul-jobs project.  This is so
we can continue to keep the upstream testing of zuul-jobs out
of the repo itself (to make for easier downstream consumption).
In the "zuul" tenant, these extra jobs and in-repo project config
will be loaded.

Change-Id: I613142d1103fdc238aa38403dbd268934efbf74d