Enable IAD opendev.org mirror
This is running afs-next branch @
579cb95f6d3a8f7db0ee361d6e8ed0c448c7a244.  This contains a number of
AFS fixes proposed for 5.3 series which we would like to sign-off as
tested in our environment.

In particular, we want to montior for messages like:

  kAFS: Volume 536871010 'mirror.opensuse' is offline

across vos release of the mirrors.  There should be changes to
fallover to alterntive servers better during release to avoid such
errors.

Prior reported fscache issues [1] are still are known problem and
being worked on.  We'll have to monitor the server; we can revert this
change after we have had sufficient testing if there are issues.

See

 https://etherpad.openstack.org/p/opendev-mirror-afs

[1] https://www.redhat.com/archives/linux-cachefs/2019-June/msg00011.html

Change-Id: I5d696907729fedd60783b13aac6537fb75c975fc