Default zuul tenant to use ansible 2.8 for jobs
After some basic testing of jobs in the zuul tenant, we don't appear to
have any issues using ansible 2.8. With that in mind, switch the default
for all jobs to ansible 2.8. We also only do this on the zuul tenant, to
ensure this tenant works properly before moving to other tenants.

Change-Id: Ifb440750c18aef96b42080cc50799e5ed8703619
Signed-off-by: Paul Belanger <pabelanger@redhat.com>