New project request: airship/porthole
This moves the Porthole project, which has been incubating in
github, into the Airship namespace.  Porthole is a project that
provides a collection of utility containers which can be used
to access various Airship and OpenStack-Helm component CLIs
(e.g. etcdctl, psql...), and can be configured for enterprise
use cases that require fine-grained RBAC controls.

Change-Id: I40c4c40292bbb9543cbcbd560731d89610f9a99d