Stop treating .gz files as special in log handling
With our switch to swift hosted logs files ended in .gz are no longer
equivalent to the filenames without the .gz. Swift expects us to be
speific. For this reason normalize on using the actual file names as
stored and reported rather than removing the suffix for human
friendliness.

Note we keep the .gz less tag value in the tags list. This is because
e-r typically uses tags to identify files rather than pure paths or
filenames.

Change-Id: I88ec2eea371fd82ebfc84a8b403cf6f192e25a5e
Related-Change: https://review.opendev.org/#/c/677236