Identify Cinder as the Block Storage Service
According to openstack/governance/projects.yaml, Cinder is the
OpenStack Block Storage service.  This patch changes the file
used to generate the specs.openstack.org index page and RSS feed
page to identify Cinder properly.  The change is only to displayed
text, URLs are not modified by this commit.

Change-Id: I2730dd82f1b6dd329a36e73150fec9c36470fc17
Partial-bug: #1841645