Add per-project Airship groups
Per discussion at the last PTG, new Airship projects have been
getting created with per-project core groups, instead of the older
airship-core group.  This change catches up existing Airship projects
to this setup, as they either have new core reviewers being added
or will in the future.  A few projects which are closely and
logically related to other projects are configured to share their
ACLs for simplicity's sake.

Note that this removes the airship.config ACL, but we do need to retain
the airship-core group (for now).

With (or ahead of) this change, please seed the existing airship-core
group as a member of all the newly created groups, to ensure
uninterrupted review/merge ability.

Change-Id: I0c054714ca9a709bd58b85570a5c415a76cacd1c