Move security site to AFS publishing
Move from promoting to static.o.o for the security site to
publish via promote to AFS.

Rename existing promote-security jobs to promote-security-static and
add it to ossa repo so
that we publish to both old and new location. The promote-security-static
job can be removed once everything works.

Create new promote-security job for publishing to AFS.

Story: 2006598
Task: 36854
Change-Id: I18404d8fa79566de1bee1ac5760b8a440e885df8