Add system-config and zuul to OpenDev tenant for jobs
This adds repos to the OpenDev tenant so that jobs defined in their
repos may be used in that tenant.  Specifically, we're looking to
run third-party-check builds of Gerrit and Zuul on changes to
upstream Gerrit repos.  The actual jobs we run there are likely to
be defined in the system-config and zuul repos.

Change-Id: I41b345246f2012d15d969524b5879c9da32b708d