Bye, bye readthedocs-failing
We introduced the readthedocs-failing template over a year ago so
that repos get updated (see [1]).
This experiment succeeded for some repos, for others it failed.

Telling teams for over a year about this problem is IMHO enough, so
let's stop this.

This change replaces the readthedocs-failing template with a template
that just builds the docs but does not publish so that we can remove the
readthedocs-failing template.

For yaql repository, it remove the job, instead change
https://review.opendev.org/685619 is proposed to import the template.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-August/132836.html

Change-Id: I722d8848b1a1877d758e30c14537f553b93da168