Remove qcow2 version 0.10 compat from our images
A long time ago in a galaxy far far away there existed the planet
hpcloud. In order to breath the atmosphere of this planet VM images had
to be compatible with the qcow2 0.10 version specification. Since then
this planet has been visited by a death star and none of our images are
expected to visit planets with this requirement.

Go ahead and stop producing VM images that can live on HPCloud.

Note that our Ubuntu and Debian images had already stopped doing this so
this should be safe. Improvements we should see as a result include
smaller images.

Change-Id: I6888b478e93021bbede4a647af990774126b0245