Bump min-ready for ubuntu-bionic
Flip the min ready values for Xenial and Bionic. This should've been
done when we did the initial transition from Xenial to Bionic but was
missed. We want to have ready nodes for the common up to date ubuntu
platform (Bionic) and don't as many on the previous release (Xenial).

Thank you to zbr for catching this after noticing that linter jobs took
a while to start.

Change-Id: I5e2c4b6a056b50077ab343535a4bbe3bc49a0acf