Add zuul/nodepool and openstack/diskimage-builder to opendev tenant
See WIP change https://review.opendev.org/696000 where we wish to
build docker images in system-config from the master branch of these
two projects.  Thus we neeed to import them as required-projects.

Somehow, zuul both reports a configuration failure and runs the jobs!
I think opendev/system-config is a weird repo that sort of lives in
openstack too.  I believe this is the cause of the failure in that
job.

Change-Id: I334edb00071031483f5a3e4b4715f773c0f261b1