Import openstacksdk and dib into zuul tenant
We are working on having our docker container builds be able to
install sibling packages.  To make nodepool jobs to build images with
dib/openstacksdk as required-projects: we need to import them.

I have removed the diskimage-builder from the full job import section,
because I do not believe we use this any more.  This was when the
nodepool functional tests were defined in dib, but they have moved to
be defined inside nodepool.  nodepool doesn't run any of the jobs
defined by diskimage-builder (in fact the opposite, dib runs the
nodepool jobs).

Change-Id: I5673adce8f92994a17e8601f6a41a597f1ba7716