Add ovn-octavia-provider driver project
This patch adds a new project under the Neutron group -
OVN Octavia Provider driver, for Neutron's
interaction with the OVN LoadBalancer for Octavia.

A few days ago the Neutron community decided to merge
Networking-ovn tree with the Neutron tree [0]. Nowadays,
the OVN Octavia provider driver is placed in networking-ovn
repository and it uses Octavia and Octavia-lib. To not
force Neutron to rely on Octavia we should export the driver
to a dedicated repository.

[0] https://review.opendev.org/#/c/658414/

Needed-By: https://review.opendev.org/#/c/697095/
Change-Id: Id4a5beb324e2fa0d119df532c24fc28503155d1b