IRC #openstack-ironic gerritbot CI failed messages
When a gerrit comment of type Verified with value -2 is posted, send a
message to the #openstack-ironic channel.

This will raise awareness of CI job failures preventing changes from
landing, and is proposed as a trial for later refinement.

Update irc_tests.py for the dependend change to pass tests.

Change-Id: I251e9e5ea760fe5da1a8fbca7cee365dd3ae563b
Depends-On: https://review.opendev.org/#/c/698089/