New project: go-redfish
This new project will host go client library for the Redfish
host management API.  It is intended to be consumable/useful
outside of Airship itself; the Airship team created this library
only because of deficiencies in other available implementations.

The folks that own the github repo from which the extant
implementation should be imported have blessed this change, and
will continue their work under the Airship umbrella.

Please add me (Matt McEuen) as an initial member of the new ACL,
and I can add additional seed core reviewers.

Thanks!

Change-Id: I0c7542461555f5a23eddf0b3d3e548a734af778e