zuul: arm64 pipeline
ARM64 testing is of interest to a number of projects.  For example:

 https://review.opendev.org/662456 : kolla
 https://review.opendev.org/676111 : diskimage-builder
 https://review.opendev.org/690798 : system-config (mirror tests)

However, we have been hesitant to put these tests into the main check
queue because with only a small number of nodes available (8
currently) it may lead to long delays merging code.

While jobs could go in the on-demand experimental pipeline, this
doesn't quite reflect the status of the jobs and nodes which are
operational.

I think that a separate queue to allow these jobs to run
automatically, but asynchronously to the regular check queue is a good
option here.  This will be a step to build a strong track-record for
the jobs without affecting the main gate, which should hopefully lead
to more interest and potentially more resources.

Change-Id: I77aac13521049496d43758121788aa44f0ecc990