Remove networking-ovs-dpdk jobs
This change removes the python jobs template
and experimental job so that they can be managed
in repo.

This jobs adds the publish-to-pypi to allow future releases
to automatically be published if we choose to make them.
networking-ovs-dpdk currently uses a branch only releases
mode as required since it primary function is to provide a devstack
plugin that is generally version independent.

Change-Id: Ie7238a742005c38e2f910615175bfe7ef2d55622