Define check-release-approval executor job
As discussed in [1], create a job that will check that a release request
has been approved by the PTL or a release liaison attached to the
deliverable.

Since this job is very short and will be rerun every time a vote is
posted, rather than consume a remote host, it will be run directly on
the executor. To that effect, the python script called only uses stdlib
or modules that are already present on the executor (requests and yaml).

[1] http://lists.openstack.org/pipermail/openstack-infra/2019-December/006556.html

Change-Id: Ibe1f0fba6ae2a459be22b33d8de4285b739a1df0