Generate PEP503 index pages for wheel mirrors
PEP503 [1] defines the simple respository API format that we should
follow for our mirrors publishing .whl files.

This loops the output directories and creates a index.html for each
project with the required tags.

We do not really have gate testing for this, so to avoid destroying
the whole mirror we write out a temporary index file first, and ignore
errors.  When this works, we will revert to index.html and do a single
manual removal of the temporary indexes.

[1] https://www.python.org/dev/peps/pep-0503/

Change-Id: I0541fb58535f27589b7e3bc1462cb083e3d12f56