copy-wheels : generate indexes version 2
Some things have become evident when generating the indexes requiring
some larger changes.

Firstly, the indexer script needs python3 on the host.  Since we're
still building CentOS 7 wheels, we need to install Python 3 from EPEL
there.

Secondly, because part of the PEP503 index page is the file hash,
reading all the files back over AFS is quite slow.  It's also quite
slow having ansible loop a task each time, which all adds up to job
timeouts.

Instead, make the indexes on the local disk before we copy the results
to AFS.  This requires copying both scripts to the host for execution
(rather than relying on "script:" ) so the wheel-copy.sh script can
call wheel-indexer.py.

While we are there, a small refactor on the wheel-indexer.py to use
os.walk() (which makes it easier to have this as a stand-alone
recursive script later, if something changes).  Also update the output
to use <ul><li> for the filenames, so it looks a little better on the
output html.

Change-Id: I85f9e132bc55fd8d33583a698e15c47665e5cf8d