Stop forcing universal wheel releases
We added the --universal flag to always be included in python release
jobs in I09cf4e64194395417c1bfa95237b04bd316c18fa to make sure projects
that did not set it in their setup.cfg would still get universal wheels.
This worked well while we had both py2 and py3 support, but now that we
have transitioned to just py3, we do not still want this to happen.

We would still like it to happen on stable releases that do still
support py2, but that really is up to the project to set the flag in
setup.cfg. Removing this from the job definition is a trade off, but
it's considered slightly better that we do not have universal wheels for
our new deliverables since not having it for stable will just mean
slightly longer install times.

Change-Id: I41de5a1f57eed0a165bcd93ab1585003c3d478fb
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>