Set root mount options on debian buster image
This is an attempt to get better fs IO performance out of our images.
Start with this on the buster image as this gives us a low impact test
if things go wrong, but also provides a system similar ish to ubuntu.

Specifically we disable barriers and set noatime as we don't care about
data reliability after crashes on our test nodes (they are single use)
and atime shouldn't matter for our CI system.

Change-Id: Ied44590c6260691a738ed105f8da0f4765b6c162