Update branch checkout and u-c.txt for Octavia
Octavia has a diskimage-builder element to install the amphora-agent.
When creating an amphora image from an Octavia stable branch, the
expectation is the amphora-agent code will match the same branch (master
or stable). However, the amphora-agent always pulls master versions of
upper-constraints.txt and of the Octavia Git repository.

This patch updates the branch name and upper constraints for the
amphora-agent in Octavia on stable branch creation.

Change-Id: I8eba64c886c187c8652f94735ca6153702203d17